fix(ci): satisfy changelog check and sdk test isolation
Rename the changelog fragment to this PR's number.

The export-path tests imported opentelemetry-exporter-otlp-proto-common, which
is not part of the sdk test environment, so they would fail collection in CI.
Move them to that exporter's own test suite, which already depends on the sdk,
and keep the sdk tests to the Resource contract itself.
